+++ vice/src/arch/unix/gui/uidatasette.c
@@ -36,13 +36,13 @@
 
 static UI_CALLBACK(ui_datasette_control)
 {
-    int command = (int)UI_MENU_CB_PARAM;
+    int command = (int)(long)UI_MENU_CB_PARAM;
     datasette_control(command);
 }
 
 static UI_CALLBACK(datasette_settings)
 {
-    int what = (int)UI_MENU_CB_PARAM;
+    int what = (int)(long)UI_MENU_CB_PARAM;
     char *prompt, *title, *resource;
     char buf[50];
     ui_button_t button;
