+++ vice/src/arch/unix/gui/uidriveiec.c
@@ -88,7 +88,7 @@ UI_CALLBACK(uidriveiec_expansion_control)
         int type;
 
         resources_get_int_sprintf("Drive%iType", &type,
-                                  (int)UI_MENU_CB_PARAM + 8);
+                                  (int)(long)UI_MENU_CB_PARAM + 8);
 
         if (drive_check_expansion(type))
             ui_menu_set_sensitive(w, 1);
@@ -105,7 +105,7 @@ UI_CALLBACK(uidriveiec_idle_method_control)
         int type;
 
         resources_get_int_sprintf("Drive%iType", &type,
-                                  (int)UI_MENU_CB_PARAM + 8);
+                                  (int)(long)UI_MENU_CB_PARAM + 8);
 
         if (drive_check_idle_method(type))
             ui_menu_set_sensitive(w, 1);
