+++ vice/src/arch/unix/gui/uijoystick2.c
@@ -41,11 +41,11 @@ static UI_CALLBACK(set_joystick_device_1)
     int tmp;
 
     if (!CHECK_MENUS) {
-        resources_set_int("JoyDevice1", (int)UI_MENU_CB_PARAM);
+        resources_set_int("JoyDevice1", (int)(long)UI_MENU_CB_PARAM);
         ui_update_menus();
     } else {
         resources_get_int("JoyDevice1", &tmp);
-        ui_menu_set_tick(w, tmp == (int)UI_MENU_CB_PARAM);
+        ui_menu_set_tick(w, tmp == (int)(long)UI_MENU_CB_PARAM);
     }
 }
 
@@ -54,11 +54,11 @@ static UI_CALLBACK(set_joystick_device_2)
     int tmp;
 
     if (!CHECK_MENUS) {
-        resources_set_int("JoyDevice2", (int)UI_MENU_CB_PARAM);
+        resources_set_int("JoyDevice2", (int)(long)UI_MENU_CB_PARAM);
         ui_update_menus();
     } else {
         resources_get_int("JoyDevice2", &tmp);
-        ui_menu_set_tick(w, tmp == (int)UI_MENU_CB_PARAM);
+        ui_menu_set_tick(w, tmp == (int)(long)UI_MENU_CB_PARAM);
     }
 }
